Move User repository to Core
