+++ src/Wallabag/ImportBundle/Consumer/AbstractConsumer.php
@@ -9,8 +9,8 @@ use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Wallabag\CoreBundle\Entity\Entry;
 use Wallabag\CoreBundle\Entity\Tag;
 use Wallabag\CoreBundle\Event\EntrySavedEvent;
+use Wallabag\CoreBundle\Repository\UserRepository;
 use Wallabag\ImportBundle\Import\AbstractImport;
-use Wallabag\UserBundle\Repository\UserRepository;
 
 abstract class AbstractConsumer
 {
