+++ tests/Wallabag/ImportBundle/Consumer/RedisEntryConsumerTest.php
@@ -7,9 +7,9 @@ use PHPUnit\Framework\TestCase;
 use Symfony\Component\EventDispatcher\EventDispatcher;
 use Wallabag\CoreBundle\Entity\Entry;
 use Wallabag\CoreBundle\Entity\User;
+use Wallabag\CoreBundle\Repository\UserRepository;
 use Wallabag\ImportBundle\Consumer\RedisEntryConsumer;
 use Wallabag\ImportBundle\Import\AbstractImport;
-use Wallabag\UserBundle\Repository\UserRepository;
 
 class RedisEntryConsumerTest extends TestCase
 {
